`ci` code under `ci` package
