+++ src/tcn/ci/pipeline/ci_info.py
@@ -1,25 +1,25 @@
 from typing import Dict, Any
-from tcn.pipeline.task import TaskBase
+from tcn.ci.pipeline.task import TaskBase
 from tcn.utils.registry import Registry
 from tcn.utils.environment import Environment
 from tcn.utils.shell import ShellScript
 
 
 @Registry.register
-class WIP(TaskBase):
+class CIInfo(TaskBase):
     def run_action(
         self,
         config: Dict[str, Any],
         env: Environment,
         metadata: Dict[str, Any],
     ):
-        # Build GEOS
-        ShellScript(name="wip").write(
-            modules=[],
-            env_to_source=[],
-            shell_commands=[
-                "srun -A j1013 -C rome --qos=4n_a100 --partition=gpu_a100 --nodes=2 --ntasks=6 --ntasks-per-node=3 --gpus-per-node=3 --sockets-per-node=2 --mem-per-gpu=40G --output=log.%t.out nvidia-smi",
-            ],
+        super().__init__(skip_metadata=True)
+        ShellScript("showquota").write(
+            [
+                "showquota",
+                "cd /discover/nobackup/gmao_ci/",
+                "du -a | cut -d/ -f2 | sort | uniq -c | sort -nr",
+            ]
         ).execute()
 
     def check(
