+++ src/tcn/ci/pipeline/templates/gpu-wrapper-slurm-mps.sh.tpl
